+++ lib/gollum-lib.rb
@@ -23,7 +23,7 @@ require File.expand_path('../gollum-lib/file_view', __FILE__)
 require File.expand_path('../gollum-lib/markup', __FILE__)
 require File.expand_path('../gollum-lib/markups', __FILE__)
 require File.expand_path('../gollum-lib/sanitization', __FILE__)
-require File.expand_path('../gollum-lib/web_sequence_diagram', __FILE__)
+require File.expand_path('../gollum-lib/filter', __FILE__)
 
 # Set ruby to UTF-8 mode
 # This is required for Ruby 1.8.7 which gollum still supports.
