+++ lib/gollum-lib/filter/render.rb
@@ -1,0 +1,18 @@
+# ~*~ encoding: utf-8 ~*~
+
+class Gollum::Filter::Render < Gollum::Filter
+  def extract(data)
+    begin
+      data = GitHub::Markup.render(@markup.name, data)
+      if data.nil?
+        raise "There was an error converting #{@markup.name} to HTML."
+      end
+    rescue Object => e
+      data = html_error("Failed to render page: #{e.message}")
+    end
+
+    data
+  end
+
+  def process(d) d; end
+end
